requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.1.1
djangorestframework-simplejwt==5.1.0
django-filter==21.1